Add ContentNav component in svelte
A new file named ContentNav.svelte was added to the lib components. This component takes in a post as input and displays navigation links based on the post headings. It also tracks users' scroll position to highlight the currently active headings. A change was also detected in the package-lock.json file.